refactor: rename simple_run_tracker.py to comprehensive_run_tracker.py
- Rename file to match the ComprehensiveRunTracker class name
- Update imports in resource.py and sqlmesh_asset_execution_utils.py
- Maintain consistency between filename and class name

This improves code organization and makes the file structure more intuitive.
